==> R/01_load_data.qmd <==
---
title: "01_load_data"
format:
  html:
    embed-resources: true
editor: visual
---

# Clear Environment

```{r}
rm(list = ls())
```

# Load libraries

```{r}
library("tidyverse")
```

# Load and save raw data

```{r}
#| message: false

# Create directory
fs::dir_create(path = '../data/_raw')

# Retrieve data if it does not exist
raw_data <- read_tsv(file = "https://hbiostat.org/data/repo/nhgh.tsv")

# Write raw data to disc
write_tsv(x = raw_data,
          file = "../data/_raw/nhgh.tsv")

```

# Create artificially split data file (for educational purposes)

```{r}

#Reading the data files into two seperate tibbles to demonstrate joining later
meta_data <- read_tsv(file = "../data/_raw/nhgh.tsv",
                      col_select = c(1:5))

measurements <- read_tsv(file = "../data/_raw/nhgh.tsv",
                         col_select = -c(2:5))

#Writing into two data files
write_tsv(x = meta_data,
          file = "../data/01_nhgh_meta_data.tsv")

write_tsv(x = measurements,
          file = "../data/01_nhgh_measurements.tsv")

```
