==> R/02_clean_data.qmd <==
---
title: "02_clean_data"
format:
  html:
    embed-resources: true
editor: visual
---

# Clear Environment

```{r}
rm(list = ls())
```

# Load libraries

```{r}
library("tidyverse")
```

# Load data

```{r}
#| message: false
meta_data <- read_tsv(file = "../data/01_nhgh_meta_data.tsv",
                      na = "NA")

measurements <- read_tsv(file = "../data/01_nhgh_measurements.tsv",
                         na = "NA")
```

# Data wrangling

```{r}
# Round age to the nearest full year
meta_data <- meta_data |>
  mutate(age = floor(age))


# Rename columns in the metadata
meta_data <- meta_data |>
  rename("id" = "seqn",
         "ethnicity" = "re")

```

Some of the income values are arbitrary categories (\> 20000, \< 20000). These values will be changed to NA.

```{r}
# Change arbitrary categories in the income values to NA
meta_data <- meta_data |> 
  mutate(income = case_when(
    income == '< 20000' ~ NA,
    income == '> 20000' ~ NA,
    .default = income
  ))
```

```{r}
# Rename columns in the measurement data 
measurements <- measurements |>
  rename("id" = "seqn",
         "medication" = "tx",
         "diagnosis" = "dx",
         "weight" = "wt",
         "height" = "ht",
         "tri_skinfold" = "tri",
         "sub_skinfold" = "sub",
         "creatinine" = "SCr")
```

# Join the data into one tibble

```{r}
diabetes_data <- 
  inner_join(x = meta_data,
             y = measurements,
             by = join_by(id))
```

# Write data to a clean datafile

```{r}
write_tsv(x = diabetes_data,
          file = "../data/02_diabetes_data_clean.tsv")
```
