==> R/03_augment.qmd <==
---
title: "03_augment"
format:
  html:
    embed-resources: true
editor: visual
---

# Load libraries

```{r}
library("tidyverse")
```


# Load data

```{r}
#| message: false
diabetes_data <- read_tsv(file = "../data/02_diabetes_data_clean.tsv",
                          na = "NA")
```

# Augment

```{r}
# Calculating "body fat percentage" and "conicity index".
# Add new variables
diabetes_data <- diabetes_data |> 
  mutate(bin_sex = case_when(
                    sex == "male" ~ 1,
                    sex == "female" ~ 0
                  ),
         bfp = 1.39*bmi+0.16*age-10.34*bin_sex-9,
         waist_m = waist/100,
         height_m = height/100,
         ci = waist_m/(0.109*sqrt(weight/(height_m))),
         dm_status = case_when(
                      diagnosis == 0 ~ 1,
                      diagnosis == 1 & medication == 0 ~ 2,
                      diagnosis == 1 & medication == 1 ~ 3
         ))
```

```{r}
# Relocate the dm_status (Disease/Medication status)
diabetes_data <- diabetes_data |> 
  relocate(dm_status, .after = diagnosis)
```

# Write data

```{r}
# Write data 
write_tsv(x = diabetes_data,
          file = "../data/03_diabetes_data_augment.tsv")
```
