==> R/05_diabetes_income_age.qmd <==
---
title: "Data Analysis I: Diabetes vs Income and Age"
format:
  html:
    embed-resources: true
editor: visual
---

### Clear environment
```{r}
rm(list = ls())
```

### Load libraries
```{r}
#| message: false
library("tidyverse")
library("patchwork")
```

### Load data
```{r}
#| message: false
diabetes_data <- read_tsv(file = "../data/03_diabetes_data_augment.tsv",
                          na = 'NA')
```

This visualisation aims to investigate whether family income (and therefore, whealthy lifestyle) is a factor or not on diabetes manifestation.
```{r}
plot1_data <- diabetes_data |>
                  mutate(income = factor(income, levels = c('[0,5000)','[5000,10000)','[10000,15000)','[15000,20000)','[20000,25000)',
                        '[25000,35000)','[35000,45000)','[45000,55000)','[55000,65000)','[65000,75000)','[75000,100000)',
                        '>= 100000'))) |>
                  drop_na() |>
                  group_by(income) |>
                  summarize(percentage = sum(diagnosis == 1)*100 / n())


plot1 <- plot1_data |>
        ggplot(mapping = aes(x = income,
                            y = percentage)
        ) +
        geom_bar(
          stat = "identity",
          color = "black",
          fill = 'wheat',
          alpha = 0.5
        ) +
        theme_minimal() +
        theme(
          text = element_text(size=5.3),
          axis.text.x = element_text(angle = 15,
                                     vjust = 1,
                                     hjust = 1),
          plot.title = element_text(hjust = 0.5),
          plot.background = element_rect(fill = "white"),
          panel.background = element_rect(fill = "white")
        ) +
        coord_cartesian(ylim = c(0, 25)) +
        labs(
          title = "Diabetes diagnosed people vs Family income classes",
          x = "Income classes",
          y = "% of diabetes diagnosed"
        )

plot1
```

This visualisation aims to investigate whether family income might be a determinant factor on the choice to pursue medications given being diagnosed with diabetes.
```{r}
plot2_data <- diabetes_data |>
                  mutate(income = factor(income, levels = c('[0,5000)','[5000,10000)','[10000,15000)','[15000,20000)','[20000,25000)',
                        '[25000,35000)','[35000,45000)','[45000,55000)','[55000,65000)','[65000,75000)','[75000,100000)',
                        '>= 100000'))) |>
                  drop_na() |>
                  filter(diagnosis==1) |>
                  group_by(income) |>
                  summarize(percentage = sum(medication == 1)*100 / n())

plot2 <- plot2_data |>
          ggplot(mapping = aes(x = income,
                              y = percentage)
          ) +
          geom_bar(
            stat = "identity",
            color = "black",
            fill = 'wheat',
            alpha = 0.5
          ) +
          theme_minimal() +
          theme(
            text = element_text(size=5.3),
            axis.text.x = element_text(angle = 15,
                                       vjust = 1,
                                       hjust = 1),
            plot.title = element_text(hjust = 0.5),
            plot.background = element_rect(fill = "white"),
            panel.background = element_rect(fill = "white")
          ) +
          coord_cartesian(ylim = c(0, 85)) +
          labs(
            title = "Medicated people vs Family income classes",
            x = "Income classes",
            y = "% of medicated people"
          )


plot2
```

###Combining the plots
```{r}
combined_plot <- plot1 / plot2
combined_plot

ggsave(combined_plot, 
       filename = "../results/05_bar_income.png",
       device = "png",
       width = 960,
       height = 1080,
       units = "px"
       )

```

This visualisation aims to investigate how the age can be a factor to take into account within being susceptible to diabetes.
```{r}
plot3_data <- diabetes_data |>
                  arrange(age) |>
                  group_by(age) |>
                  summarize(percentage = sum(diagnosis == 1)*100 / n())

plot3 <- plot3_data |>
          ggplot(mapping = aes(x = age,
                              y = percentage)
                ) +
                geom_bar(
                  stat = "identity", 
                  color = "black",
                  fill = "wheat",
                  alpha = 0.5
                ) +
                scale_x_continuous(
                  breaks = seq(12, 80, 2)
                ) +
                theme_minimal() +
                theme(
                  text = element_text(size = 5.3),
                  plot.title = element_text(hjust = 0.5),
                  plot.background = element_rect(fill = "white"),
                  panel.background = element_rect(fill = "white")
                ) +
                coord_cartesian(ylim = c(0, 60)) +
                labs(
                  title = "Diagnosed people vs Age",
                  x = "Age",
                  y = "% of diabetes diagnosed"
                )

ggsave(plot3, 
       filename = "../results/05_bar_age.png",
       device = "png",
       width = 960,
       height = 1080,
       units = "px"
       )

plot3
```
